refactor: 디렉토리 삭제 및 파일 위치 이동
