+++ back/src/main/java/col/carrot/back/user/userlogin/domain/UserRepository.java
@@ -1,4 +1,4 @@
-package col.carrot.back.user.userlogin;
+package col.carrot.back.user.userlogin.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
